applied review feedback: - domain exception classes TheOnlyFolderRemoveException, ForeignFolderRemoveException - move domain logic from application layer - repository check methods isUserHasFolder, isUserHasMoreThanOneFolder - extend folder fixtures to test foreign folder remove - cest requestShouldReturn400IfRemoveForeignFolder

+++ src/Domain/Service/FolderService.php
@@ -5,6 +5,8 @@ declare(strict_types=1);
 namespace App\Domain\Service;
 
 use App\Domain\Entity\ValueObject\Id;
+use App\Domain\Exception\ForeignFolderRemoveException;
+use App\Domain\Exception\TheOnlyFolderRemoveException;
 use App\Domain\Repository\FolderRepositoryInterface;
 
 final class FolderService implements FolderServiceInterface
@@ -16,13 +18,19 @@ final class FolderService implements FolderServiceInterface
         $this->folderRepository = $folderRepository;
     }
 
-    public function userHaveTheOnlyFolder(Id $userId): bool
+    private function user(Id $userId): bool
     {
-        return 1 === count($this->folderRepository->getByUserId($userId));
+        return count($this->folderRepository->getByUserId($userId)) > 1;
     }
 
-    public function delete(Id $folderId): void
+    public function delete(Id $userId, Id $folderId): void
     {
+        if (!$this->folderRepository->isUserHasFolder($userId, $folderId)) {
+            throw new ForeignFolderRemoveException();
+        }
+        if (!$this->folderRepository->isUserHasMoreThanOneFolder($userId)) {
+            throw new TheOnlyFolderRemoveException();
+        }
         $this->folderRepository->delete($folderId);
     }
 }
